Extended range of DateTime64 to years 1698 - 2378
* Added TimeZoneImpl that mimics DateLUTImpl static interface,
  but has multiple such LUTs (some are loaded on-demand) and routes calls
  to corresponding LUT depending on date/time_t argument value.
* Updated DateLUTImpl to properly support values close to epoch start (1970-01-01 00:00).
* Reduced resolution of DateLUTImpl::Values::time_at_offset_change
  to multiple of 15-minutes to allow storing 64-bits of time_t in DateLUTImpl::Value.
* Minor performance updates to DateLUTImpl when building month LUT
  by skipping non-start-of-month days.
* Added GlobalDayNum (typedef to Int32) to allow values outside of 1970-2105 range.
* Fixed extractTimeZoneFromFunctionArguments to work correctly with DateTime64.
* TimeZoneImpl is now properly integrated into all (or most) DateTime64-related functions.
* New unit-tests and stateless integration tests for DateTime and DateTime64.

And lots of other minor things.

+++ src/DataTypes/DataTypeDateTime64.h
@@ -3,6 +3,7 @@
 #include <Core/Types.h>
 #include <DataTypes/DataTypeDateTime.h>
 #include <DataTypes/DataTypeDecimalBase.h>
+#include <common/TimeZone.h>
 
 class DateLUTImpl;
 
@@ -103,7 +104,7 @@ public:
         else
         {
             const auto components = DecimalUtils::splitWithScaleMultiplier(t, scale_multiplier);
-            return transform->execute(static_cast<UInt32>(components.whole), std::forward<Args>(args)...);
+            return transform->execute(static_cast<Int64>(components.whole), std::forward<Args>(args)...);
         }
     }
